backend/requirements.txt
faker
imgkit
num2words
pandas
pillow
playwright
python-dotenv
supabase
clerk-backend-api
fastapi
uvicorn
sqlalchemy